updated return type of methods and test cases to accept decimal values

+++ src/main/java/ibat/calc/MyMath.java
@@ -3,19 +3,19 @@ package ibat.calc;
 
 public class MyMath {
 	
-	public int testAdd(int a, int b) {
+	public double testAdd(int a, int b) {
 		return a+b;
 	}
 	
-	public int testSubtract(int a, int b) {
+	public double testSubtract(int a, int b) {
 		return a-b;
 	}
 	
-	public int testMultiply(int a, int b) {
+	public double testMultiply(int a, int b) {
 		return a*b;
 	}
 	
-	public int testDivide(int a, int b) {
+	public double testDivide(int a, int b) {
 		return a/b;
 	}
 }
